Gather configuration checks in THOP main function

+++ src/thop/server/config.hh
@@ -15,7 +15,7 @@ struct Config {
     HeapArray<const char *> mco_stay_directories;
     HeapArray<const char *> mco_stay_filenames;
 
-    uint16_t port = 8888;
+    int port = 8888;
 
     BlockAllocator str_alloc { Kibibytes(16) };
 };
